+++ testapps/GHIssues/src/main/java/ghissues/BaseController.java
@@ -2,7 +2,6 @@ package ghissues;
 
 import act.handler.NoReturnValueAdvice;
 import act.util.LogSupport;
-import org.osgl.aaa.NoAuthenticate;
 import org.osgl.aaa.NoAuthentication;
 
 @NoReturnValueAdvice